JAMES-1901 Move ES metrics reporter to a dedicated project
