+++ server/container/metrics/metrics-dropwizard/pom.xml
@@ -57,10 +57,6 @@
             <groupId>javax.inject</groupId>
             <artifactId>javax.inject</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.elasticsearch</groupId>
-            <artifactId>metrics-elasticsearch-reporter</artifactId>
-        </dependency>
     </dependencies>
 
 </project>
